Pulled Description and IsEnabled up to JobItem.
