--- FileRenamer.Core/Jobs/IJobItem.cs
@@ -1,9 +1,0 @@
-﻿using FileRenamer.Core.Serialization;
-
-
-namespace FileRenamer.Core.Jobs;
-
-public interface IJobItem: IXmlSerializableAsync
-{
-	void Run(JobTarget target, JobContext context);
-}
